+++ feature/s3/transfermanager/api_op_GetObject.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"math"
 	"strconv"
 	"strings"
 	"sync"
@@ -645,12 +644,12 @@ func (g *getter) get(ctx context.Context) (out *GetObjectOutput, err error) {
 			return nil, err
 		}
 
-		partsCount := int32(math.Max(float64(aws.ToInt32(out.PartsCount)), 1))
-		partSize := int64(math.Max(float64(aws.ToInt64(out.ContentLength)), 1))
-		sectionParts := int32(math.Max(1, float64(g.options.GetBufferSize/partSize)))
+		partsCount := max(aws.ToInt32(out.PartsCount), 1)
+		partSize := max(aws.ToInt64(out.ContentLength), 1)
+		sectionParts := int32(max(1, g.options.GetBufferSize/partSize))
 		capacity := sectionParts
 		g.r.setPartSize(partSize)
-		g.r.setCapacity(int32(math.Min(float64(capacity), float64(partsCount))))
+		g.r.setCapacity(min(capacity, partsCount))
 		g.r.setPartsCount(partsCount)
 
 		ch := make(chan getChunk, g.options.Concurrency)
@@ -666,7 +665,7 @@ func (g *getter) get(ctx context.Context) (out *GetObjectOutput, err error) {
 			}
 
 			if g.r.getRead() == capacity {
-				capacity = int32(math.Min(float64(capacity+sectionParts), float64(partsCount)))
+				capacity = min(capacity+sectionParts, partsCount)
 				g.r.setCapacity(capacity)
 			}
 
@@ -710,10 +709,10 @@ func (g *getter) get(ctx context.Context) (out *GetObjectOutput, err error) {
 		total := g.totalBytes - g.offset
 
 		partsCount := int32((total-1)/g.options.PartSizeBytes + 1)
-		sectionParts := int32(math.Max(1, float64(g.options.GetBufferSize/g.options.PartSizeBytes)))
+		sectionParts := int32(max(1, g.options.GetBufferSize/g.options.PartSizeBytes))
 		capacity := sectionParts
 		g.r.setPartSize(g.options.PartSizeBytes)
-		g.r.setCapacity(int32(math.Min(float64(capacity), float64(partsCount))))
+		g.r.setCapacity(min(capacity, partsCount))
 		g.r.setPartsCount(partsCount)
 
 		ch := make(chan getChunk, g.options.Concurrency)
@@ -729,7 +728,7 @@ func (g *getter) get(ctx context.Context) (out *GetObjectOutput, err error) {
 			}
 
 			if g.r.getRead() == capacity {
-				capacity = int32(math.Min(float64(capacity+sectionParts), float64(partsCount)))
+				capacity = min(capacity+sectionParts, partsCount)
 				g.r.setCapacity(capacity)
 			}
 
@@ -786,7 +785,7 @@ func (g *getter) singleDownload(ctx context.Context, clientOptions ...func(*s3.O
 		return nil, err
 	}
 
-	g.r.setPartSize(int64(math.Max(1, float64(aws.ToInt64(out.ContentLength)))))
+	g.r.setPartSize(max(1, aws.ToInt64(out.ContentLength)))
 	g.r.setCapacity(1)
 	g.r.setPartsCount(1)
 	g.r.ch <- outChunk{body: bytes.NewReader(buf), length: aws.ToInt64(out.ContentLength)}
@@ -881,7 +880,7 @@ func (g *getter) getDownloadRange() (int64, int64) {
 // byteRange returns an HTTP Byte-Range header value that should be used by the
 // client to request a chunk range.
 func (g *getter) byteRange() string {
-	return fmt.Sprintf("bytes=%d-%d", g.pos, int64(math.Min(float64(g.totalBytes-1), float64(g.pos+g.options.PartSizeBytes-1))))
+	return fmt.Sprintf("bytes=%d-%d", g.pos, min(g.totalBytes-1, g.pos+g.options.PartSizeBytes-1))
 }
 
 func (g *getter) getErr() error {
@@ -914,193 +913,4 @@ type outChunk struct {
 
 	length int64
 	cur    int64
-}
-
-// ConcurrentReader receives object parts from working goroutines, composes those chunks in order and read
-// to user's buffer. ConcurrentReader limits the max number of chunks it could receive and read at the same
-// time so getter won't send following parts' request to s3 until user reads all current chunks, which avoids
-// too much memory consumption when caching large object parts
-type ConcurrentReader struct {
-	ch  chan outChunk
-	buf map[int32]*outChunk
-
-	partsCount int32
-	capacity   int32
-	count      int32
-	read       int32
-
-	written  int64
-	partSize int64
-
-	m sync.Mutex
-
-	err error
-}
-
-// NewConcurrentReader returns a ConcurrentReader used in GetObject input
-func NewConcurrentReader() *ConcurrentReader {
-	return &ConcurrentReader{
-		buf:      make(map[int32]*outChunk),
-		partSize: 1, // just a placeholder value
-	}
-}
-
-// Read implements io.Reader to compose object parts in order and read to p.
-// It will receive up to r.capacity chunks, read them to p if any chunk index
-// fits into p scope, otherwise it will buffer those chunks and read them in
-// following calls
-func (r *ConcurrentReader) Read(p []byte) (int, error) {
-	if cap(p) == 0 {
-		return 0, nil
-	}
-
-	var written int
-
-	for r.count < r.getCapacity() {
-		if e := r.getErr(); e != nil && e != io.EOF {
-			r.written += int64(written)
-			r.clean()
-			return written, r.getErr()
-		}
-		if written >= cap(p) {
-			r.written += int64(written)
-			return written, r.getErr()
-		}
-
-		oc, ok := <-r.ch
-		if !ok {
-			r.written += int64(written)
-			return written, r.getErr()
-		}
-
-		r.count++
-		index := r.getPartSize()*int64(oc.index) - r.written
-
-		if index < int64(cap(p)) {
-			n, err := oc.body.Read(p[index:])
-			oc.cur += int64(n)
-			written += n
-			if err != nil && err != io.EOF {
-				r.setErr(err)
-				r.clean()
-				r.written += int64(written)
-				return written, r.getErr()
-			}
-		}
-		if oc.cur < oc.length {
-			r.buf[oc.index] = &oc
-		} else {
-			r.incrRead(1)
-			if r.getRead() >= r.partsCount {
-				r.setErr(io.EOF)
-			}
-		}
-	}
-
-	partSize := r.getPartSize()
-	minIndex := int32(r.written / partSize)
-	maxIndex := int32(math.Min(float64(((r.written + int64(cap(p)) - 1) / partSize)), float64(r.getCapacity()-1)))
-	for i := minIndex; i <= maxIndex; i++ {
-		if e := r.getErr(); e != nil && e != io.EOF {
-			r.written += int64(written)
-			r.clean()
-			return written, r.getErr()
-		}
-
-		c, ok := r.buf[i]
-		if ok {
-			index := int64(i)*partSize + c.cur - r.written
-			n, err := c.body.Read(p[index:])
-			c.cur += int64(n)
-			written += n
-			if err != nil && err != io.EOF {
-				r.setErr(err)
-				r.clean()
-				r.written += int64(written)
-				return written, r.getErr()
-			}
-			if c.cur >= c.length {
-				r.incrRead(1)
-				delete(r.buf, i)
-				if r.getRead() >= r.partsCount {
-					r.setErr(io.EOF)
-				}
-			}
-		}
-	}
-
-	r.written += int64(written)
-	return written, r.getErr()
-}
-
-func (r *ConcurrentReader) setPartSize(n int64) {
-	r.m.Lock()
-	defer r.m.Unlock()
-
-	r.partSize = n
-}
-
-func (r *ConcurrentReader) getPartSize() int64 {
-	r.m.Lock()
-	defer r.m.Unlock()
-
-	return r.partSize
-}
-
-func (r *ConcurrentReader) setCapacity(n int32) {
-	r.m.Lock()
-	defer r.m.Unlock()
-
-	r.capacity = n
-}
-
-func (r *ConcurrentReader) getCapacity() int32 {
-	r.m.Lock()
-	defer r.m.Unlock()
-
-	return r.capacity
-}
-
-func (r *ConcurrentReader) setPartsCount(n int32) {
-	r.m.Lock()
-	defer r.m.Unlock()
-
-	r.partsCount = n
-}
-
-func (r *ConcurrentReader) incrRead(n int32) {
-	r.m.Lock()
-	defer r.m.Unlock()
-
-	r.read += n
-}
-
-func (r *ConcurrentReader) getRead() int32 {
-	r.m.Lock()
-	defer r.m.Unlock()
-
-	return r.read
-}
-
-func (r *ConcurrentReader) setErr(err error) {
-	r.m.Lock()
-	defer r.m.Unlock()
-
-	r.err = err
-}
-
-func (r *ConcurrentReader) getErr() error {
-	r.m.Lock()
-	defer r.m.Unlock()
-
-	return r.err
-}
-
-func (r *ConcurrentReader) clean() {
-	for {
-		_, ok := <-r.ch
-		if !ok {
-			break
-		}
-	}
 }
